+++ src/components/upload/Upload.vue
@@ -25,7 +25,6 @@
             v-bind="$attrs"
             :multiple="multiple"
             :disabled="disabled"
-            @click="upload"
             @change="onFileChange">
     </label>
 </template>
@@ -48,6 +47,10 @@
             type: {
                 type: String,
                 default: 'is-primary'
+            },
+            native: {
+                type: Boolean,
+                default: false
             }
         },
         data() {
@@ -61,23 +64,18 @@
             /**
              * When v-model is changed:
              *   1. Set internal value.
-             *   2. If it's invalid, validate again.
+             *   2. Reset input value if array is empty
+             *   3. If it's invalid, validate again.
              */
             value(value) {
                 this.newValue = value
+                if (this.newValue.length === 0) {
+                    this.$refs.input.value = null
+                }
                 !this.isValid && !this.dragDrop && this.checkHtml5Validity()
             }
         },
         methods: {
-
-            /**
-             * Reset input value
-             */
-            upload() {
-                if (!this.disabled && !this.loading) {
-                    this.$refs.input.value = null
-                }
-            },
 
             /**
              * Listen change event on input type 'file',
@@ -85,7 +83,6 @@
              */
             onFileChange(event) {
                 if (this.disabled || this.loading) return
-
                 if (this.dragDrop) {
                     this.updateDragDropFocus(false)
                 }
@@ -102,8 +99,11 @@
                         } else {
                             this.newValue = []
                         }
+                    } else {
+                        if (this.native) {
+                            this.newValue = []
+                        }
                     }
-
                     for (let i = 0; i < value.length; i++) {
                         this.newValue.push(value[i])
                     }